MovieFragment implementation. ViewModel used to control the view state, and provide data to be used on the fragment.
Most of the code went to the MovieFragment, where it'll show the current status of the movie request, if the network request went good or not. Based on that, when the response return essential informations about the movies, the fragment will show data on a layout structure based on TodoMovies4 movie screen. The fragment has a recyclerView that'll show data from similar movies by the SimilarMoviesAdapter.